editable zindex

+++ src/armature_window.rs
@@ -120,7 +120,7 @@ pub fn new_bone(bones: &mut Vec<Bone>, parent_id: i32) -> (Bone, usize) {
         scale: Vec2 { x: 1., y: 1. },
         tex_idx: -1,
         pivot: Vec2::new(0., 0.),
-        zindex: bones.len() as i32,
+        zindex: bones.len() as f32,
         ..Default::default()
     };
     if parent_id == -1 {
